pyproject.toml
[project]
name = "telco-churn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feature-engine>=1.9.3",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.6",
    "pandas>=2.3.3",
    "pydantic>=2.11.9",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
